move Reset to the new class
